fix(Component): changed according to comments
