+++ src/components/Main/JuiciestSection/JuiciestSection.tsx
@@ -2,8 +2,9 @@ import { Box, Button, Flex, Image, Text } from '@chakra-ui/react';
 import { Link, useNavigate } from 'react-router';
 
 import { CategoryTags } from '~/components/CategoryPage/TabComponent/CategoryTags/CategoryTags';
-import { Category } from '~/query/services/category-api';
+import { Category } from '~/query/services/category-api.type';
 import { useGetRecipesQuery } from '~/query/services/recipe-api';
+import { checkAndNavigate } from '~/utils/checkAndNavigate';
 import { getFullMediaUrl } from '~/utils/getFullMediaUrl';
 
 import bookmarkHeart from './../../../assets/actionBar/BookmarkHeart.svg';
@@ -23,30 +24,17 @@ export const JuiciestSection = ({ categoryData }: JuiciestSectionProps) => {
         sortOrder: 'desc',
     });
 
-    const dataCategories = categoryData?.filter((item) => item.subCategories);
-    const dataSubCategories = categoryData?.filter((item) => !item.subCategories);
-
     const handleGetRecipe = (recipeId: string, categoriesIds: string[]) => {
-        if (!dataSubCategories || !dataCategories) {
+        const { condition, matchedCategory, matchedSubcategory } = checkAndNavigate({
+            categoriesIds,
+            categoryData,
+        });
+        if (condition) {
             navigate('/error-page');
             return;
         }
-        const matchedSubcategory = dataSubCategories.find((sub) => categoriesIds.includes(sub._id));
-        if (!matchedSubcategory) {
-            navigate('/error-page');
-            return;
-        }
-        const matchedCategory = dataCategories.find(
-            (cat) => cat._id === matchedSubcategory.rootCategoryId,
-        );
-        if (!matchedCategory) {
-            navigate('/error-page');
-            return;
-        }
-        navigate(`/${matchedCategory.category}/${matchedSubcategory.category}/${recipeId}`);
+        navigate(`/${matchedCategory?.category}/${matchedSubcategory?.category}/${recipeId}`);
     };
-
-    console.log(data?.data);
 
     return (
         <Flex
