+++ src/constants/testIds.ts
@@ -1,0 +1,3 @@
+export const TEST_IDS = {
+    filterMenuButtonCategory: 'filter-menu-button-категория',
+};
