+++ src/query/services/category-api.type.ts
@@ -1,0 +1,16 @@
+export type Category = {
+    _id: string;
+    title: string;
+    category: string;
+    icon: string;
+    description: string;
+    subCategories: SubCategory[];
+    rootCategoryId: string;
+};
+
+export type SubCategory = {
+    _id: string;
+    title: string;
+    category: string;
+    rootCategoryId: string;
+};
